refactor: restructure db access
